Huge diff: + implemented string comparison using DP and + refactored server communication;

+++ integriscreen/app/src/main/java/com/example/integriscreen/UIElement.java
@@ -13,7 +13,7 @@ public class UIElement {
     public String defaultVal;   // default value of the UI element
     public String currentVal;   // current value of the UI element extracted through OCR
     public boolean found;       // true if the UI element is found on the screen through OCR
-    public double lastUpdated;  // track the time of last update
+    public long lastUpdated;  // track the time of last update
     public Boolean dirty;       // true if the element is changed simultanuesly with another one
 
 //    public ArrayList<> traces;  // list all changes for an UI element
